+++ BitmovinAnalyticsCollector/Classes/stateMachines/StateMachineDelegate.swift
@@ -3,12 +3,12 @@ import Foundation
 
 protocol StateMachineDelegate: class {
     func stateMachineDidExitSetup(_ stateMachine: StateMachine)
-    func stateMachine(_ stateMachine: StateMachine, didExitBufferingWithDuration duration: Int)
+    func stateMachine(_ stateMachine: StateMachine, didExitBufferingWithDuration duration: Int64)
     func stateMachineDidEnterError(_ stateMachine: StateMachine)
-    func stateMachine(_ stateMachine: StateMachine, didExitPlayingWithDuration duration: Int)
-    func stateMachine(_ stateMachine: StateMachine, didExitPauseWithDuration duration: Int)
+    func stateMachine(_ stateMachine: StateMachine, didExitPlayingWithDuration duration: Int64)
+    func stateMachine(_ stateMachine: StateMachine, didExitPauseWithDuration duration: Int64)
     func stateMachineDidQualityChange(_ stateMachine: StateMachine)
-    func stateMachine(_ stateMachine: StateMachine, didExitSeekingWithDuration duration: Int, destinationPlayerState: PlayerState)
-    func stateMachine(_ stateMachine: StateMachine, didHeartbeatWithDuration duration: Int)
-    func stateMachine(_ stateMachine: StateMachine, didStartupWithDuration duration: Int)
+    func stateMachine(_ stateMachine: StateMachine, didExitSeekingWithDuration duration: Int64, destinationPlayerState: PlayerState)
+    func stateMachine(_ stateMachine: StateMachine, didHeartbeatWithDuration duration: Int64)
+    func stateMachine(_ stateMachine: StateMachine, didStartupWithDuration duration: Int64)
 }
